Moved facade-specific classes out of core
